Move some functions to the utils provider from SendDashboardProperties.
