+++ provider/utils.py
@@ -1,3 +1,4 @@
+import re
 
 S3_DATE_FORMAT = '%Y%m%d%H%M%S'
 
@@ -6,3 +7,12 @@ def pad_msid(msid):
 
 def pad_volume(volume):
     return '{:02d}'.format(int(volume))
+
+def tidy_whitespace(string):
+    string = re.sub('\n', ' ', string)
+    string = re.sub(' +', ' ', string)
+    string = string.strip()
+    return string
+
+def article_status(is_poa):
+    return 'POA' if is_poa else 'VOR'
